Change Dependency from JSON::XS to JSON
JSON::XS depends on libc6, the GNU C Library, a binary which must be compiled.
JSON on the other hand selects the best available implementation:

This module is a thin wrapper for JSON::XS-compatible modules with a few additional features. All the backend modules convert a Perl data structure to a JSON text and vice versa. This module uses JSON::XS by default, and when JSON::XS is not available, falls back on JSON::PP, which is in the Perl core since 5.14. If JSON::PP is not available either, this module then falls back on JSON::backportPP (which is actually JSON::PP in a different .pm file) bundled in the same distribution as this module. You can also explicitly specify to use Cpanel::JSON::XS, a fork of JSON::XS by Reini Urban.

https://metacpan.org/pod/JSON#DESCRIPTION

Signed-off-by: Mikko Johannes Koivunalho <mikko.koivunalho@iki.fi>